Add ability to get backtraces from guest
Summary:
Allows the tracer to fetch a stack trace from the guest. Note that this will not demangle symbol names nor try to find the location of a symbol.

I'm using libunwind because it deals with all the complexity of walking the stack. There is no native Rust library to do this unfortunately (only FFI shim libraries). `libunwind-ptrace` is the only library that I'm aware of that can get remote stack traces.

Reviewed By: wangbj

Differential Revision: D32889571

fbshipit-source-id: e1e0430435c6c7f36e1cb312723a598b6b5bb2a1