Add file and line info to backtraces
Summary:
Adds file and line number resolution to backtraces. This also adds a 1GB limit on the amount of in-memory symbols at any given time. If the internal symbol cache exceeds this, the oldest entries will be discarded until we're under that limit again. We may need to tweak this and make it more configurable in the future.

Follow-up: Auxiliary debug files, such as `.debug` files, are not supported. This should be easy to support by reading the `gnu_debuglink` section and finding the `.debug` file.

Reviewed By: johnhurt

Differential Revision: D33086623

fbshipit-source-id: 43112e3bdd135a1a43ca75c10756cc8ec101c9b8