Command getter for TracerBuilder
Summary:
Adds a way to fetch the command provided to `TracerBuilder`.

We're running into a situation where we're constructing a TracerBuilder and using it much later, but we want to debug and print the command that we're tracing when we execute the traced program.

You can alternatively avoid this by storing the necessary info alongside in an adjacent struct, but it feels a little silly to have to carry around a separate struct just for debugging.

Reviewed By: jasonwhite

Differential Revision: D34231169

fbshipit-source-id: ef40cb396f5530ee1f4324694ce5d9122dfe281f