Fix bug in check_for_xen_pmi_bug
Summary: The `count` wasn't getting initialized properly, so it was possible in theory that the check could have been failing. I noticed this while attempting to convert this chunk of `llvm_asm!()` to the new `asm!()` syntax. Also added documentation explaining what this code is doing because it's not easy to read.

Reviewed By: johnhurt

Differential Revision: D34542806

fbshipit-source-id: 0f79b9d36757756a1cf76e98830f9cc76f06336c