minor: print args to syscall
Summary:
If we're going to print a trace message when we inject the syscall, we might as well show the raw args.

The specific use case I used this for was making sure when I am changing args between the Detcore and reverie layers, that they are changing in the expected way.  Detcore already prints log messages for the syscalls and arguments it receives from the guest.  This lets us see the actual arguments that are passed to the kernel.

Reviewed By: jasonwhite

Differential Revision: D35804814

fbshipit-source-id: 2fc5a1b526326b554d4f015bfd0cb41d0dfc4ab2