Use SysnoSet in Subscription
Summary: The `syscalls` crate now has the ability to create a statically-allocated bitset of syscalls. This gets rid of our usage of the dynamically-allocated `BitVec` and no longer makes assumptions about the size and offset of the syscall table. These assumptions are wrong for some architectures where syscall numbers start at 4096 instead of 0.

Reviewed By: johnhurt

Differential Revision: D37890556

fbshipit-source-id: 8d0d51e93c1a09c7390fd3898cea7955ef786e96