Bump SKID_MARGIN_RCBS from 60 to 70
Summary:
For AMD hosts I'm seeing issues with trying to run `mkfs.ext3` at large sizes using Antlir. We can raise it slightly to deal with this issue.

If it becomes too much of a problem for Intel hosts, we can always break out compilation for the two processor types

Reviewed By: jasonwhite

Differential Revision: D56271985

fbshipit-source-id: 1af3a007363a241cfa1c3e12585d6f837b274d19