Update gitoxide to version with lenient tag parsing
Summary:
ignore-conflict-markers

A while ago, I made the parsing of tags in gitoxide tolerate a missing
timestamp.

We have now seen this situation in two repos:
* aosp/platform/external/brctl
* aosp/platform/external/iproute2

Update all gitoxide crates to their latest version (equivalent to upstream main)
to propagate that fix to us.

I had to list all gitoxide crates explicitely to avoid issues with conflicting requirements
for any crates in the dependency chain.

The commit I'm pointing at in my fork of gitoxide is litterally `main` plus a version update
to propagate the fix throughout the gix ecosystem.

NOTE: I had to slightly edit the integration test to make the failure
representative of the incorrect parsing, but see the updated
integration test for proof that the latest version can parse a tag that's
missing a timestamp.

Reviewed By: RajivTS, singhsrb

Differential Revision: D63771328

fbshipit-source-id: fdbc9d762bacf48c7b17615181371cd92a4a921f