add support for tracing openat2 to reverie
Summary:
# This Diff
Add support for openat2 which was listed as missing in reverie.
Added a test as well to validate

# This Stack

We use reverie in our buckification tool `makeover` and I noticed that some tools (compiler.par) was failing only on CI.
My best guess is that when we run PAR files on CI they try to read a Metadata file and will do so via `openat2`:

https://www.internalfb.com/code/fbsource/[75d708271359]/fbcode/common/base/pid_info/Util.cpp?lines=1011
&
https://www.internalfb.com/code/fbsource/[75d708271359]/fbcode/common/base/pid_info/Util.cpp?lines=95

Since we were missing tracing for these syscalls, our tool would panic that subsequent syscalls were attempting to use an `fd` it had no clue about.

```
2025-12-10T19:46:31.407868Z  WARN tracer::process_tracer: Untracked FD 20 (points to Some("/METADATA")) in read syscall
```

This stack seeks to add support for `openat2` to makeover and reverie.

1. < Step 1>
2. **Add support for openat2 to reverie**
3. < Step 3 >

Differential Revision: D88970559

fbshipit-source-id: 6b7375b58262bac5a986afcffd4d158f68e63f4c